refactor: move auth persistence owners
