+++ packages/shared-server/mod.ts
@@ -9,9 +9,9 @@ export * from './app-data/AppDataRepository.ts';
 export * from './app-data/RallarServerAppData.ts';
 export * from './rallar-system/middleware/RallarMiddleware.ts';
 export * from './rallar-system/cache-repositories.ts';
-export * from './rallar-system/repositories/AuthSessionRepository.ts';
+export * from './rallar-system/auth/persistence/auth-session-repository.ts';
 export { hashAuthSecret } from './rallar-system/auth/credentials/hash-auth-secret.ts';
-export * from './rallar-system/repositories/AuthUserRepository.ts';
+export * from './rallar-system/auth/persistence/auth-user-repository.ts';
 export * from './rallar-system/client-state/persistence/client-state-repository.ts';
 export * from './rallar-system/repositories/GroupStateRepository.ts';
 export * from './rallar-system/repositories/GroupTopologyConfigRepository.ts';
