+++ packages/shared-server/rallar-system/group-state/group-mutation-authority.ts
@@ -3,8 +3,8 @@ import type { StateScope } from '@shared/api/state-types.ts';
 import { NonRetryableException } from '@shared/queuebox/DequeueResourceEntryController.ts';
 
 import type { AuthSessionRepository } from '../repositories/AuthSessionRepository.ts';
-import type { PersistedAuthSession } from '../repositories/auth-persistence-contracts.ts';
-import type { IssuedAuthSession } from '../repositories/auth-session-types.ts';
+import type { PersistedAuthSession } from '../auth/persistence/auth-persistence-contracts.ts';
+import type { IssuedAuthSession } from '../auth/persistence/auth-session-types.ts';
 import { authSessionProofSecret } from '../auth/sessions/auth-session-proof-secret.ts';
 import {
   canonicalJson,
