--- packages/shared-server/rallar-system/services/auth-state-codecs.ts
@@ -1,2 +1,0 @@
-export { decodeAuthMutationCommand } from '../auth/mutation/decode-auth-mutation-command.ts';
-export { decodeAuthMutationResult } from '../auth/mutation/decode-auth-mutation-result.ts';
