+++ packages/tests/shared-server/app-auth-persistence-inbox.test.ts
@@ -331,10 +331,10 @@ describe('AppAuthInboxService architecture', () => {
     it('removes auth and AL domain-lock escape hatches from production', () => {
         const sources = [
             'packages/shared-server/rallar-system/services/auth-login-service.ts',
-            'packages/shared-server/rallar-system/repositories/AuthSessionRepository.ts',
-            'packages/shared-server/rallar-system/repositories/auth-session-persistence.ts',
-            'packages/shared-server/rallar-system/repositories/auth-ticket-persistence.ts',
-            'packages/shared-server/rallar-system/repositories/auth-legacy-compatibility.ts',
+            'packages/shared-server/rallar-system/auth/persistence/auth-session-repository.ts',
+            'packages/shared-server/rallar-system/auth/persistence/auth-session-persistence.ts',
+            'packages/shared-server/rallar-system/auth/persistence/auth-ticket-persistence.ts',
+            'packages/shared-server/rallar-system/auth/persistence/auth-legacy-compatibility.ts',
             'packages/shared-server/postgres/al-runtime/PSqlInboundAdmissionBackend.ts',
             'packages/shared-server/postgres/al-runtime/PSqlOutboundAdmissionBackend.ts',
             'packages/shared-server/postgres/runtime-state/PSqlRuntimeStateRepository.ts',
